# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce REST API built using Go Lang, PSQL, REST API following Clean Architecture. Hosted with Ngnix, AWS EC2 and RDS",
        "title": "Ecommerce REST API",
        "contact": {
            "name": "Nadeem Fahad",
            "url": "https://github.com/Nadeem1815",
            "email": "nadeemf408@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/admin/dashboard": {
            "get": {
                "description": "Admin Can access dashboard and view details of recoding orders,products etc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "DashBoard",
                "operationId": "dash_board",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/download": {
            "get": {
                "description": "Admin can download total sales report in csv.format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Sales Report",
                "operationId": "sales_report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/find_userid/:user_id": {
            "get": {
                "description": "Admin Can Find All Register Users details find By user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Find User By Id",
                "operationId": "find-user-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user to be fetched",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/list_all_user": {
            "get": {
                "description": "Admin Can List All Register Users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List All Users",
                "operationId": "list-all-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/login/email": {
            "post": {
                "description": "Admin login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Login",
                "operationId": "admin-login",
                "parameters": [
                    {
                        "description": "Admin login credentials",
                        "name": "admin_credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.AdminLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "description": "Logs out a logged-in admin from the E-commerce web api admin panel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin Logout",
                "operationId": "admin-logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/register": {
            "post": {
                "description": "New Admin Registration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin SignUp",
                "operationId": "admin-signup",
                "parameters": [
                    {
                        "description": "Register Admin ",
                        "name": "newAdmin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_domain.Admin"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_nadeem1815_premium-watch_pkg_domain.Admin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_nadeem1815_premium-watch_pkg_utils_model.AdminLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_nadeem1815_premium-watch_pkg_utils_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}